mainFolder/crowd_management/requirements.txt
streamlit==1.31.1
opencv-python==4.9.0.80
ultralytics==8.1.2
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
seaborn==0.13.2
matplotlib==3.8.2
playsound==1.3.0 